+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased][unreleased]
 
+### Added
+- Update last used app icons immediately after sending a new app
+
 <a id="1_59_2"></a>
 
 ## [1.59.2] - 2025-06-25